+++ R/val_labels.R
@@ -230,23 +230,56 @@ val_label.data.frame <- function(x, v, prefixed = FALSE) {
 #' @param .data a data frame
 #' @param ... name-value pairs of value labels (see examples)
 #' @note
-#'   \code{add_val_labels} could be used with \code{dplyr} (see examples).
+#'   \code{set_value_labels} and \code{add_value_labels} could be used with \code{dplyr}.
+#'   While \code{set_value_labels} will replace the list of value labels, \code{add_value_labels}
+#'   will update that list (see examples).
 #' @return
-#'  \code{add_val_labels} will return a named vector.
+#'  \code{set_value_labels} and \code{add_value_labels} will return an updated copy of \code{.data}.
 #' @examples
 #' if (require(dplyr)) {
+#'   # setting value labels
 #'   df <- data_frame(s1 = c("M", "M", "F"), s2 = c(1, 1, 2)) %>%
-#'     add_val_labels(s1 = c(Male = "M", Female = "F"), s2 = c(Yes = 1, No = 2))
+#'     set_value_labels(s1 = c(Male = "M", Female = "F"), s2 = c(Yes = 1, No = 2))
+#'   val_labels(df)
+#'
+#'   # updating value labels
+#'   df <- df %>% add_value_labels(s2 = c(Unknown = 9))
+#'   val_labels(df)
+#'
+#'   # removing a value labels
+#'   df <- df %>% add_value_labels(s2 = c(NULL = 9))
 #'   val_labels(df)
 #' }
 #' @export
-add_val_labels <- function(.data, ...) {
+set_value_labels <- function(.data, ...) {
   values <- list(...)
   if (!all(names(values) %in% names(.data)))
     stop("some variables not found in .data")
 
   for (v in names(values))
     val_labels(.data[[v]]) <- values[[v]]
+
+  .data
+}
+
+#' @rdname val_labels
+#' @export
+add_value_labels <- function(.data, ...) {
+  values <- list(...)
+  if (!all(names(values) %in% names(.data)))
+    stop("some variables not found in .data")
+
+  for(v in values)
+    if (is.null(names(v)) | any(names(v) == ""))
+      stop("all arguments should be named vectors")
+
+  for (v in names(values))
+    for (l in names(values[[v]])) {
+      if (l == "NULL")
+        val_label(.data[[v]], values[[v]][[l]]) <- NULL
+      else
+        val_label(.data[[v]], values[[v]][[l]]) <- l
+    }
 
   .data
 }