+++ R/var_label.R
@@ -74,3 +74,35 @@ var_label.data.frame <- function(x) {
   for (var in names(value)) var_label(x[[var]]) <- value[[var]]
   x
 }
+
+
+#' @rdname var_label
+#' @param .data a data frame
+#' @param ... name-value pairs of variable labels (see examples)
+#' @note
+#'   \code{set_variable_labels} could be used with \code{dplyr}.
+#' @return
+#'  \code{set_variable_labels} will return an updated copy of \code{.data}.
+#' @examples
+#' if (require(dplyr)) {
+#'   # adding some variable labels
+#'   df <- data_frame(s1 = c("M", "M", "F"), s2 = c(1, 1, 2)) %>%
+#'     set_variable_labels(s1 = "Sex", s2 = "Yes or No?")
+#'   var_label(df)
+#'
+#'   # removing a variable label
+#'   df <- df %>% set_variable_labels(s2 = NULL)
+#'   var_label(df$s2)
+#' }
+#' @export
+set_variable_labels <- function(.data, ...) {
+  values <- list(...)
+  if (!all(names(values) %in% names(.data)))
+    stop("some variables not found in .data")
+
+  for (v in names(values))
+    var_label(.data[[v]]) <- values[[v]]
+
+  .data
+}
+